Update generate.py
We add a function that can choose which file we will use to create and new sentence with a rotatory switch with 12 positions.
For this, we ask to the Arduino to send values of different analog port to the computer. Then we read this values with python and if one of them is equal to 0 during 2s we choose a specific files, depending on which port respect this condition.